+++ ca/django_ca/management/commands/edit_ca.py
@@ -11,15 +11,18 @@
 # You should have received a copy of the GNU General Public License along with django-ca.  If not,
 # see <http://www.gnu.org/licenses/>.
 
+"""Management command to edit a certificate authority via the command line."""
+
 from ...extensions import IssuerAlternativeName
 from ..base import BaseCommand
 from ..base import CertificateAuthorityDetailMixin
 
 
-class Command(BaseCommand, CertificateAuthorityDetailMixin):
+class Command(BaseCommand, CertificateAuthorityDetailMixin):  # pylint: disable=missing-class-docstring
     help = 'Edit a certificate authority.'
 
     def add_arguments(self, parser):
+        self.add_general_args(parser, default=None)
         self.add_ca(parser, 'ca', allow_disabled=True)
         self.add_ca_args(parser)
 
@@ -29,11 +32,11 @@ class Command(BaseCommand, CertificateAuthorityDetailMixin):
         group.add_argument('--disable', action='store_false', dest='enabled',
                            help='Disable the certificate authority.')
 
-    def handle(self, ca, **options):
+    def handle(self, ca, **options):  # pylint: disable=arguments-differ
         if options['issuer_url'] is not None:
             ca.issuer_url = options['issuer_url']
         if options[IssuerAlternativeName.key]:
-            ca.issuer_alt_name = options[IssuerAlternativeName.key]
+            ca.issuer_alt_name = ','.join(options[IssuerAlternativeName.key])
         if options['ocsp_url'] is not None:
             ca.ocsp_url = options['ocsp_url']
         if options['crl_url'] is not None:
@@ -41,5 +44,12 @@ class Command(BaseCommand, CertificateAuthorityDetailMixin):
 
         if options['enabled'] is not None:
             ca.enabled = options['enabled']
+
+        if options['caa'] is not None:
+            ca.caa_identity = options['caa']
+        if options['website'] is not None:
+            ca.website = options['website']
+        if options['tos'] is not None:
+            ca.terms_of_service = options['tos']
 
         ca.save()